+++ crates/store/src/lib.rs
@@ -10,6 +10,9 @@ pub mod state;
 #[cfg(feature = "rocksdb")]
 pub use accounts::PersistentAccountTree;
 pub use accounts::{AccountTreeWithHistory, HistoricalError, InMemoryAccountTree};
+pub use db::Db;
+pub use db::models::conv::SqlTypeConvert;
+pub use errors::DatabaseError;
 pub use genesis::GenesisState;
 pub use server::block_prover_client::BlockProver;
 pub use server::{DataDirectory, Store};