+++ crates/validator/src/server/mod.rs
@@ -1,41 +1,31 @@
 use std::net::SocketAddr;
-use std::num::NonZeroUsize;
+use std::path::PathBuf;
 use std::sync::Arc;
 use std::time::Duration;
 
 use anyhow::Context;
+use miden_node_db::Db;
 use miden_node_proto::generated::validator::api_server;
 use miden_node_proto::generated::{self as proto};
 use miden_node_proto_build::validator_api_descriptor;
 use miden_node_utils::ErrorReport;
-use miden_node_utils::lru_cache::LruCache;
 use miden_node_utils::panic::catch_panic_layer_fn;
 use miden_node_utils::tracing::OpenTelemetrySpanExt;
 use miden_node_utils::tracing::grpc::grpc_trace_fn;
 use miden_protocol::block::{BlockSigner, ProposedBlock};
-use miden_protocol::transaction::{
-    ProvenTransaction,
-    TransactionHeader,
-    TransactionId,
-    TransactionInputs,
-};
+use miden_protocol::transaction::{ProvenTransaction, TransactionInputs};
 use miden_tx::utils::{Deserializable, Serializable};
 use tokio::net::TcpListener;
 use tokio_stream::wrappers::TcpListenerStream;
 use tonic::Status;
 use tower_http::catch_panic::CatchPanicLayer;
 use tower_http::trace::TraceLayer;
-use tracing::{Instrument, info_span};
+use tracing::{info_span, instrument};
 
 use crate::COMPONENT;
 use crate::block_validation::validate_block;
+use crate::db::{insert_transaction, load};
 use crate::tx_validation::validate_transaction;
-
-/// Number of transactions to keep in the validated transactions cache.
-const NUM_VALIDATED_TRANSACTIONS: NonZeroUsize = NonZeroUsize::new(10000).unwrap();
-
-/// A type alias for a LRU cache that stores validated transactions.
-pub type ValidatedTransactions = LruCache<TransactionId, TransactionHeader>;
 
 // VALIDATOR
 // ================================================================================
@@ -53,6 +43,9 @@ pub struct Validator<S> {
 
     /// The signer used to sign blocks.
     pub signer: S,
+
+    /// The data directory for the validator component's database files.
+    pub data_directory: PathBuf,
 }
 
 impl<S: BlockSigner + Send + Sync + 'static> Validator<S> {
@@ -62,6 +55,11 @@ impl<S: BlockSigner + Send + Sync + 'static> Validator<S> {
     /// encountered.
     pub async fn serve(self) -> anyhow::Result<()> {
         tracing::info!(target: COMPONENT, endpoint=?self.address, "Initializing server");
+
+        // Initialize database connection.
+        let db = load(self.data_directory.join("validator.sqlite3"))
+            .await
+            .context("failed to initialize validator database")?;
 
         let listener = TcpListener::bind(self.address)
             .await
@@ -86,7 +84,7 @@ impl<S: BlockSigner + Send + Sync + 'static> Validator<S> {
             .layer(CatchPanicLayer::custom(catch_panic_layer_fn))
             .layer(TraceLayer::new_for_grpc().make_span_with(grpc_trace_fn))
             .timeout(self.grpc_timeout)
-            .add_service(api_server::ApiServer::new(ValidatorServer::new(self.signer)))
+            .add_service(api_server::ApiServer::new(ValidatorServer::new(self.signer, db)))
             .add_service(reflection_service)
             .add_service(reflection_service_alpha)
             .serve_with_incoming(TcpListenerStream::new(listener))
@@ -103,14 +101,12 @@ impl<S: BlockSigner + Send + Sync + 'static> Validator<S> {
 /// Implements the gRPC API for the validator.
 struct ValidatorServer<S> {
     signer: S,
-    validated_transactions: Arc<ValidatedTransactions>,
+    db: Arc<Db>,
 }
 
 impl<S> ValidatorServer<S> {
-    fn new(signer: S) -> Self {
-        let validated_transactions =
-            Arc::new(ValidatedTransactions::new(NUM_VALIDATED_TRANSACTIONS));
-        Self { signer, validated_transactions }
+    fn new(signer: S, db: Db) -> Self {
+        Self { signer, db: db.into() }
     }
 }
 
@@ -128,6 +124,7 @@ impl<S: BlockSigner + Send + Sync + 'static> api_server::Api for ValidatorServer
     }
 
     /// Receives a proven transaction, then validates and stores it.
+    #[instrument(target = COMPONENT, skip_all, err)]
     async fn submit_proven_transaction(
         &self,
         request: tonic::Request<proto::transaction::ProvenTransaction>,
@@ -150,17 +147,17 @@ impl<S: BlockSigner + Send + Sync + 'static> api_server::Api for ValidatorServer
         tracing::Span::current().set_attribute("transaction.id", tx.id());
 
         // Validate the transaction.
-        let validated_tx_header = validate_transaction(tx, inputs).await.map_err(|err| {
+        let tx_info = validate_transaction(tx, inputs).await.map_err(|err| {
             Status::invalid_argument(err.as_report_context("Invalid transaction"))
         })?;
 
-        // Register the validated transaction.
-        let tx_id = validated_tx_header.id();
-        self.validated_transactions
-            .put(tx_id, validated_tx_header)
-            .instrument(info_span!("validated_txs.insert"))
-            .await;
-
+        // Store the validated transaction.
+        self.db
+            .transact("insert_transaction", move |conn| insert_transaction(conn, &tx_info))
+            .await
+            .map_err(|err| {
+                Status::internal(err.as_report_context("Failed to insert transaction"))
+            })?;
         Ok(tonic::Response::new(()))
     }
 
@@ -181,11 +178,12 @@ impl<S: BlockSigner + Send + Sync + 'static> api_server::Api for ValidatorServer
 
         // Validate the block.
         let signature =
-            validate_block(proposed_block, &self.signer, self.validated_transactions.clone())
-                .await
-                .map_err(|err| {
-                    tonic::Status::invalid_argument(format!("Failed to validate block: {err}",))
-                })?;
+            validate_block(proposed_block, &self.signer, &self.db).await.map_err(|err| {
+                tonic::Status::invalid_argument(format!(
+                    "Failed to validate block: {}",
+                    err.as_report()
+                ))
+            })?;
 
         // Send the signature.
         info_span!("serialize").in_scope(|| {